Update encoding profiles 3215 (#15)
* Update audio encoder to native aac and prevent unnecessary upscaling

* Update min Expression Syntax to Work With Matterhorn

Testing the previously committed file raised some issues: Matterhorn didn't like the shell-style single quotes, and the more proper usage of the expression syntax for FFmpeg is the comma. In addition, Matterhorn needs two back slash characters to escape the comma or it barfs. This file should now play just fine in Matterhorn.

* Increase maximum height for derivatives